refactor: PostLikeService 클래스 수정
